refactor: extract plugins to cms/plugins directory
